feat: prepare URL and webview
