+++ p2p_wallet/Scenes/Sell/SellViewModel.swift
@@ -6,7 +6,6 @@ class SellViewModel: BaseViewModel, ObservableObject {
     
     // MARK: - Dependencies
     
-    @Injected private var userWalletManager: UserWalletManager
     // TODO: - Use resolver instead
     private let actionService = SellActionServiceMock()
     private let dataService = SellDataServiceMock()
@@ -14,38 +13,23 @@ class SellViewModel: BaseViewModel, ObservableObject {
     // MARK: - Properties
 
 //    @Published var state: SellStateMachine.State
-    @Published private(set) var subscene: SellSubScene?
-    
-    private let externalTransactionId: String = UUID().uuidString
+    private var subscene = PassthroughSubject<SellSubScene?, Never>()
+    var subscenePublisher: AnyPublisher<SellSubScene?, Never> {
+        subscene.eraseToAnyPublisher()
+    }
 
     // MARK: - Actions
     
     func openMoonPayWebView(
         quoteCurrencyCode: String,
-        baseCurrencyAmount: Double
+        baseCurrencyAmount: Double,
+        externalTransactionId: String
     ) throws {
-        #if !RELEASE
-        let endpoint = String.secretConfig("MOONPAY_STAGING_SELL_ENDPOINT")!
-        let apiKey = String.secretConfig("MOONPAY_STAGING_API_KEY")!
-        #else
-        let endpoint = String.secretConfig("MOONPAY_PRODUCTION_SELL_ENDPOINT")!
-        let apiKey = String.secretConfig("MOONPAY_PRODUCTION_API_KEY")!
-        #endif
-        
-        var components = URLComponents(string: endpoint)!
-        components.queryItems = [
-            .init(name: "apiKey", value: apiKey),
-            .init(name: "baseCurrencyCode", value: "SOL"),
-            .init(name: "refundWalletAddress", value: userWalletManager.wallet?.account.publicKey.base58EncodedString),
-            .init(name: "quoteCurrencyCode", value: quoteCurrencyCode),
-            .init(name: "baseCurrencyAmount", value: baseCurrencyAmount.toString()),
-            .init(name: "externalTransactionId", value: externalTransactionId)
-        ]
-        
-        guard let url = components.url else {
-            throw SellError.invalidURL
-        }
-        
-        subscene = .moonpayWebpage(url: url)
+        let url = try actionService.createSellURL(
+            quoteCurrencyCode: quoteCurrencyCode,
+            baseCurrencyAmount: baseCurrencyAmount,
+            externalTransactionId: externalTransactionId
+        )
+        subscene.send(.moonpayWebpage(url: url))
     }
 }
